# backend/internal/api/openapi.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package api

import (
	"time"

	"lumbercalc/backend/internal/store"
)

const (
	BearerScopes bearerContextKey = "bearer.Scopes"
)

// CreateDesignRequest defines model for CreateDesignRequest.
type CreateDesignRequest struct {
	Description *string `json:"description,omitempty"`

	// Document A design at schema version 1. Lengths are inches. Rotation is Euler XYZ in degrees.
	Document *DesignDocument `json:"document,omitempty"`

	// Name The server removes leading and trailing whitespace. A blank result is rejected.
	Name string `json:"name"`
}

// Design defines model for Design.
type Design struct {
	Description string `json:"description"`

	// Document A design at schema version 1. Lengths are inches. Rotation is Euler XYZ in degrees.
	Document  DesignDocument `json:"document"`
	Id        string         `json:"id"`
	Name      string         `json:"name"`
	UpdatedAt time.Time      `json:"updatedAt"`
	Version   int            `json:"version"`
}

// DesignDocument A design at schema version 1. Lengths are inches. Rotation is Euler XYZ in degrees.
type DesignDocument = store.Document

// DesignSummary defines model for DesignSummary.
type DesignSummary struct {
	Description string    `json:"description"`
	Id          string    `json:"id"`
	Name        string    `json:"name"`
	UpdatedAt   time.Time `json:"updatedAt"`
	Version     int       `json:"version"`
}

// Me defines model for Me.
type Me struct {
	ClerkUserId string `json:"clerk_user_id"`
	DisplayName string `json:"display_name"`
	UserId      string `json:"user_id"`
}

// ReplaceDesignRequest defines model for ReplaceDesignRequest.
type ReplaceDesignRequest struct {
	Description *string `json:"description,omitempty"`

	// Document A design at schema version 1. Lengths are inches. Rotation is Euler XYZ in degrees.
	Document DesignDocument `json:"document"`

	// Name The server removes leading and trailing whitespace. A blank result is rejected.
	Name string `json:"name"`
}

// bearerContextKey is the context key for bearer security scheme
type bearerContextKey string

// CreateDesignJSONRequestBody defines body for CreateDesign for application/json ContentType.
type CreateDesignJSONRequestBody = CreateDesignRequest

// ReplaceDesignJSONRequestBody defines body for ReplaceDesign for application/json ContentType.
type ReplaceDesignJSONRequestBody = ReplaceDesignRequest