# backend/internal/store/internal/q/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package q

import (
	"encoding/json"

	"github.com/jackc/pgx/v5/pgtype"
)

type Design struct {
	ID                  pgtype.UUID
	OwnerID             pgtype.UUID
	Name                string
	Description         string
	SchemaVersion       int32
	Document            json.RawMessage
	Version             int32
	CopiedFromDesignID  pgtype.UUID
	CopiedFromPatternID pgtype.UUID
	CreatedAt           pgtype.Timestamptz
	UpdatedAt           pgtype.Timestamptz
	DeletedAt           pgtype.Timestamptz
}

type DesignMaterialUsage struct {
	DesignID   pgtype.UUID
	MaterialID pgtype.UUID
}

type Material struct {
	ID                 pgtype.UUID
	OwnerID            pgtype.UUID
	Name               string
	Species            pgtype.Text
	Grade              pgtype.Text
	NominalThicknessIn pgtype.Numeric
	NominalWidthIn     pgtype.Numeric
	ActualThicknessIn  pgtype.Numeric
	ActualWidthIn      pgtype.Numeric
	RoughThicknessIn   pgtype.Numeric
	RoughWidthIn       pgtype.Numeric
	CreatedAt          pgtype.Timestamptz
	UpdatedAt          pgtype.Timestamptz
}

type OptimizationRun struct {
	ID             pgtype.UUID
	OwnerID        pgtype.UUID
	DesignID       pgtype.UUID
	InputHash      string
	Status         string
	Result         []byte
	TotalCostCents pgtype.Int4
	ErrorCode      pgtype.Text
	ErrorDetail    []byte
	CreatedAt      pgtype.Timestamptz
	FinishedAt     pgtype.Timestamptz
}

type Pattern struct {
	ID             pgtype.UUID
	SourceDesignID pgtype.UUID
	PublisherID    pgtype.UUID
	Name           string
	Description    string
	SchemaVersion  int32
	Document       []byte
	IsPublic       bool
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
}

type Price struct {
	ID         pgtype.UUID
	OwnerID    pgtype.UUID
	MaterialID pgtype.UUID
	LengthIn   pgtype.Numeric
	PriceCents int32
	CreatedAt  pgtype.Timestamptz
	UpdatedAt  pgtype.Timestamptz
}

type User struct {
	ID          pgtype.UUID
	ClerkUserID string
	CreatedAt   pgtype.Timestamptz
}

type UserSetting struct {
	UserID                pgtype.UUID
	DefaultUseKerf        bool
	DefaultKerfIn         pgtype.Numeric
	DefaultDimensionBasis string
	UpdatedAt             pgtype.Timestamptz
}