# backend/openapi.yaml
openapi: 3.0.3
info:
  title: LumberCalc API
  version: 0.0.0
  description: >
    Paths under /v1 require a bearer session.
    This includes a path that has no operation.
    A request under /v1 with no session returns 401.
    A request under /v1 with a session and no operation returns 404.
    /healthz and /readyz do not require a session.
servers:
  - url: /
tags:
  - name: probes
  - name: v1
security:
  - bearer: []
paths:
  /healthz:
    get:
      operationId: getHealth
      tags: [probes]
      security: []
      responses:
        "200":
          description: The process is up.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "500":
          description: The process panicked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /readyz:
    get:
      operationId: getReadiness
      tags: [probes]
      security: []
      responses:
        "200":
          description: Every probe succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyOk"
        "503":
          description: A probe failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyFailed"
        "500":
          description: The process panicked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/me:
    get:
      operationId: getMe
      tags: [v1]
      responses:
        "200":
          description: The signed-in user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: The request has no valid session.
          headers:
            WWW-Authenticate:
              required: true
              schema:
                type: string
                enum: [Bearer]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: The session check or the user row failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/designs:
    post:
      operationId: createDesign
      tags: [v1]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDesignRequest"
      responses:
        "201":
          description: The design was created.
          headers:
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Design"
        "400":
          description: The name or the document is not valid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: The request has no valid session.
          headers:
            WWW-Authenticate:
              required: true
              schema:
                type: string
                enum: [Bearer]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The document names a material that does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: The design was not saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listDesigns
      tags: [v1]
      responses:
        "200":
          description: The caller's live designs, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DesignSummary"
        "401":
          description: The request has no valid session.
          headers:
            WWW-Authenticate:
              required: true
              schema:
                type: string
                enum: [Bearer]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: The designs were not loaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/designs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getDesign
      tags: [v1]
      responses:
        "200":
          description: The design, including its document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Design"
        "401":
          description: The request has no valid session.
          headers:
            WWW-Authenticate:
              required: true
              schema:
                type: string
                enum: [Bearer]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The caller has no such design.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: The design was not loaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: replaceDesign
      tags: [v1]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceDesignRequest"
      responses:
        "200":
          description: The design was replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Design"
        "400":
          description: The name or the document is not valid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: The request has no valid session.
          headers:
            WWW-Authenticate:
              required: true
              schema:
                type: string
                enum: [Bearer]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The caller has no such design.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The document names a material that does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: The design was not saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteDesign
      tags: [v1]
      responses:
        "204":
          description: The design was deleted.
        "401":
          description: The request has no valid session.
          headers:
            WWW-Authenticate:
              required: true
              schema:
                type: string
                enum: [Bearer]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The caller has no such design.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: The design was not deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
    Health:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ReadyOk:
      type: object
      additionalProperties: false
      required: [status, checks]
      properties:
        status:
          type: string
          enum: [ready]
        checks:
          type: array
          items:
            $ref: "#/components/schemas/ReadinessCheck"
    ReadyFailed:
      type: object
      additionalProperties: false
      required: [status, checks]
      properties:
        status:
          type: string
          enum: [not_ready]
        checks:
          type: array
          items:
            $ref: "#/components/schemas/ReadinessCheck"
    ReadinessCheck:
      type: object
      additionalProperties: false
      required: [name, ok]
      properties:
        name:
          type: string
          minLength: 1
        ok:
          type: boolean
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          minLength: 1
    Me:
      type: object
      additionalProperties: false
      required: [user_id, clerk_user_id, display_name]
      properties:
        user_id:
          type: string
          minLength: 1
        clerk_user_id:
          type: string
          minLength: 1
          maxLength: 64
        display_name:
          type: string
          minLength: 1
    CreateDesignRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          description: The server removes leading and trailing whitespace. A blank result is rejected.
        description:
          type: string
        document:
          $ref: "#/components/schemas/DesignDocument"
    ReplaceDesignRequest:
      type: object
      additionalProperties: false
      required: [name, document]
      properties:
        name:
          type: string
          minLength: 1
          description: The server removes leading and trailing whitespace. A blank result is rejected.
        description:
          type: string
        document:
          $ref: "#/components/schemas/DesignDocument"
    DesignSummary:
      type: object
      additionalProperties: false
      required: [id, name, description, version, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
        name:
          type: string
          minLength: 1
        description:
          type: string
        version:
          type: integer
          minimum: 1
        updatedAt:
          type: string
          format: date-time
    Design:
      type: object
      additionalProperties: false
      required: [id, name, description, version, updatedAt, document]
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
        name:
          type: string
          minLength: 1
        description:
          type: string
        version:
          type: integer
          minimum: 1
        updatedAt:
          type: string
          format: date-time
        document:
          $ref: "#/components/schemas/DesignDocument"
    DesignDocument:
      x-go-type: store.Document
      x-go-type-import:
        path: lumbercalc/backend/internal/store
      description: A design at schema version 1. Lengths are inches. Rotation is Euler XYZ in degrees.
      type: object
      additionalProperties: false
      required: [schemaVersion, units, pieces, connections]
      properties:
        schemaVersion:
          type: integer
          enum: [1]
        units:
          type: string
          enum: [in]
        pieces:
          type: array
          items:
            $ref: "#/components/schemas/DesignPiece"
        connections:
          type: array
          items:
            $ref: "#/components/schemas/DesignConnection"
    DesignPiece:
      type: object
      additionalProperties: false
      required: [id, name, materialId, lengthIn, continuity, transform]
      properties:
        id:
          $ref: "#/components/schemas/DocumentUuid"
        name:
          type: string
        materialId:
          $ref: "#/components/schemas/DocumentUuid"
        roughCut:
          type: boolean
          default: false
        lengthIn:
          type: number
          description: Exact inch length. The numeric value stays exact. An exponent form can come back as a plain decimal.
        continuity:
          $ref: "#/components/schemas/DesignContinuity"
        transform:
          $ref: "#/components/schemas/DesignTransform"
    DesignContinuity:
      type: object
      additionalProperties: false
      required: [x, y, z]
      description: One flag per local axis. True means that axis stays one board. False means the solver may splice.
      properties:
        x:
          type: boolean
        y:
          type: boolean
        z:
          type: boolean
    DesignTransform:
      type: object
      additionalProperties: false
      required: [positionIn, rotationDeg]
      properties:
        positionIn:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: number
          description: Inches, in axis order x, y, z.
        rotationDeg:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: number
          description: Euler angles in degrees, applied in X, then Y, then Z order.
    DesignConnection:
      type: object
      additionalProperties: false
      required: [id, pieceAId, pieceBId, joinType, faceA, faceB, wasteIn]
      properties:
        id:
          type: string
          minLength: 1
        pieceAId:
          $ref: "#/components/schemas/DocumentUuid"
        pieceBId:
          $ref: "#/components/schemas/DocumentUuid"
        joinType:
          type: string
          enum: [butt, lap, fastener]
        faceA:
          type: string
          minLength: 1
        faceB:
          type: string
          minLength: 1
        wasteIn:
          type: number
          description: Exact inches. The numeric value stays exact. An exponent form can come back as a plain decimal.
    DocumentUuid:
      description: A client-generated UUID. Unique among piece ids in one document. It is not a database key.
      type: string
      format: uuid